+++ R/rsprite2-package.R
@@ -13,7 +13,7 @@ rSprite.maxDeltaLoopsLower <- 20000
 rSprite.maxDeltaLoopsUpper <- 1000000
 rSprite.maxDupLoops <- 20
 
-rSprite.dust <- 1e-12 #To account for rounding errors
+rSprite.dust <- 1e-10 #To account for floating point issues
 rSprite.huge <- 1e15 #Should this not be Inf?
 
 #' @keywords internal